Add intial version of synth-02
The synth alternates between two different tones when it receives a
MIDI noteon message. The tones are generated from a 2000 sample cycle
that is then downsampled to create a cycle at the desired frequence.

The downsampled waves are generated once, at the beginning of the
program, and there are discontinuities between the beginning and the
end of the waveforms, which result in pops.

The next step is to see if new windows of the waveform can be
generated quickly enough to provide a continuous tone.